fix(reivew): hit majority of notes
Removed asset buttons
cleaned up translations
Renamed OpenMintAssetModal to MintAssetButton
