+++ src/components/designer/taro/actions/index.ts
@@ -1,3 +1,2 @@
-export { default as AssetButtons } from './AssetButtons';
 export { default as MintAssetModal } from './MintAssetModal';
-export { default as OpenMintAssetModal } from './OpenMintAssetModal';
+export { default as MintAssetButton } from './MintAssetButton';
