+++ src/i18n/locales/en-US.json
@@ -271,13 +271,9 @@
   "cmps.designer.taro.ConnectTab.filePaths": "File Paths",
   "cmps.designer.taro.ConnectTab.hexStrings": "HEX",
   "cmps.designer.taro.ConnectTab.base64Strings": "Base64",
-  "cmps.designer.taro.ActionsTab.MintAsset.title": "Minting",
-  "cmps.designer.taro.ActionsTab.MintAsset.mint": "Mint",
-  "cmps.designer.taro.ActionsTab.MintAsset.send": "Send",
-  "cmps.designer.taro.ActionsTab.MintAsset.mintAssetHelp": "Mint a Taro Asset",
-  "cmps.designer.taro.ActionsTab.MintAsset.sendAssetHelp": "Send a Taro Asset",
+  "cmps.designer.taro.ActionsTab.MintAssetButton.title": "Mint Asset",
+  "cmps.designer.taro.ActionsTab.MintAssetButton.mint": "Mint",
   "cmps.designer.taro.actions.MintAssetModal.title": "Mint an asset for {{name}}",
-  "cmps.designer.taro.actions.MintAssetModal.type": "Type",
   "cmps.designer.taro.actions.MintAssetModal.okBtn": "Mint",
   "cmps.designer.taro.actions.MintAssetModal.cancelBtn": "Cancel",
   "cmps.designer.taro.actions.MintAssetModal.assetType": "Asset Type",
@@ -291,8 +287,6 @@
   "cmps.designer.taro.actions.MintAssetModal.skipBatch": "Skip batch to mint asset immediately",
   "cmps.designer.taro.actions.MintAssetModal.amount": "Amount",
   "cmps.designer.taro.actions.MintAssetModal.amountPlaceholder": "1",
-  "cmps.designer.taro.actions.MintAssetModal.normalAmountHelp": "The target supply of the minted asset",
-  "cmps.designer.taro.actions.MintAssetModal.collectibleAmountHelp": "Only one collectible can be minted at a time",
   "cmps.designer.taro.actions.MintAssetModal.metaData": "Meta Data",
   "cmps.designer.taro.actions.MintAssetModal.metaDataPlaceholder": "Optional",
   "cmps.designer.taro.actions.MintAssetModal.mintSuccess": "Successfully minted asset: '{{name}}' with batchkey: {{batchKey}}",
