+++ src/store/models/modals.ts
@@ -1,4 +1,3 @@
-//import { actions } from '@mrblenny/react-flow-chart';
 import { Action, action, Thunk, thunk } from 'easy-peasy';
 import { StoreInjections } from 'types';
 import { RootModel } from './';
